Old maven compiler configuration and plugin.
These have been floating around my personal repository for a number of
years, and have been intermittently updated over time.  The plugin
adds a number of features — most specifically, searching the class
path to find imports and trying to intelligently place them in the
top of a java source when you enter ``\mi'' over a classname.

The compiler format has required a lot of massaging over the years,
and will probably need more because maven's compiler output is so
grossly verbose.